Dev adi

+++ app/Http/Controllers/Stable/StableController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Stable;
 
 use App\Http\Controllers\Controller;
 use App\Http\Traits\MediaUploadingTrait;
+use App\Mail\StableAdminSendSubmitApproval;
 use App\Models\City;
 use App\Models\District;
 use App\Models\Province;
@@ -11,6 +12,9 @@ use Illuminate\Http\Request;
 
 // load models
 use App\Models\Stable;
+use App\Models\User;
+use App\Models\Coach;
+use App\Models\Horse;
 use App\Models\Village;
 use App\Notifications\StableRegisteredToStableOwner;
 use Carbon\Carbon;
@@ -133,12 +137,28 @@ class StableController extends Controller
         $stable->district_id        = $request->district;
         $stable->village_id         = $request->village;
         $stable->owner              = $request->owner;
-        $stable->manager            = $request->manager;
+        $stable->manager            = $request->manager;        
+        $stable->facilities         = $request->facilities;    
         $stable->capacity_of_stable = $request->capacity_of_stable;
         $stable->capacity_of_arena  = $request->capacity_of_arena;
-        $stable->number_of_coach    = $request->number_of_coach;
-        $stable->facilities         = $request->facilities;
+        $stable->number_of_coach    = $request->number_of_coach;    
+        
+        $coachNum1 = Coach::where('stable_id', $stable->id)->get()->count();
+        $coachNum2 = $stable->number_of_coach; 
+        if($coachNum1 > $coachNum2)
+        {
+            Alert::error('Error', 'Number of coach cannot smaller than coach data')->persistent(true)->autoClose(3600);
+            return redirect()->back();
+        }
 
+        $horseNum1      = Horse::where('stable_id', $stable->id)->get()->count();
+        $stableCapacity = $stable->capacity_of_stable;                
+        if($horseNum1 > $stableCapacity)
+        {
+            Alert::error('Error', 'Capacity of stable full')->persistent(true)->autoClose(3600);
+            return redirect()->back();
+        }
+        
         if ($request->logo) {
             // delete old logo
             File::delete($stable->logo);
@@ -171,5 +191,36 @@ class StableController extends Controller
 
         Alert::error('Stable key not match.')->persistent(true)->autoClose(3600);
         return redirect()->route('stable.stable_key.confirm');
+    }
+
+    public function submitApproval($id)
+    {
+        $data = Stable::find($id);
+
+        $users1 = User::whereHas("roles", function($q){
+                            $q->where("name", "app-owner"); 
+                        })->get();
+
+        $users2 = User::whereHas("roles", function($q){
+                            $q->where("name", "app-admin"); 
+                        })->get();
+        
+        foreach($users1 as $user)
+        {
+            $send = User::where('id', $user->id)->first();
+            $send->notify(new StableAdminSendSubmitApproval($data));
+        }
+
+        foreach($users2 as $user)
+        {
+            $send = User::where('id', $user->id)->first();
+            $send->notify(new StableAdminSendSubmitApproval($data));
+        }
+
+        Stable::where('id', $data->id)->update([
+            'approval_status' => 'Need Approval'
+        ]);
+        Alert::success($data->name.' Submit Approval Sent', 'Success.')->persistent(true)->autoClose(3600);
+        return redirect()->back();
     }
 }
